web: move QueryBuilder to its own search subdir
The web/src/search/input dir was getting cluttered.
