+++ web/src/SourcegraphWebApp.scss
@@ -220,6 +220,7 @@ body,
 @import './savedSearches/index';
 @import './search/input/ScopePage';
 @import './search/input/SearchPage';
+@import './search/queryBuilder/index';
 @import './search/results/SearchResults';
 @import './settings/SettingsFile';
 @import './site-admin/SiteInitPage';
