+++ web/src/search/input/SearchPage.scss
@@ -2,8 +2,6 @@
 @import '../QuickLinks';
 @import './SearchScopes';
 @import './QueryInput';
-@import './QueryBuilder';
-@import './QueryBuilderInputRow';
 @import './InfoDropdown';
 @import './interactive/InteractiveModeInput';
