+++ web/src/search/input/SearchPage.tsx
@@ -14,7 +14,7 @@ import { ThemePreferenceProps } from '../../theme'
 import { limitString } from '../../util'
 import { submitSearch, QueryState } from '../helpers'
 import { QuickLinks } from '../QuickLinks'
-import { QueryBuilder } from './QueryBuilder'
+import { QueryBuilder } from '../queryBuilder/QueryBuilder'
 import { QueryInput } from './QueryInput'
 import { SearchButton } from './SearchButton'
 import { SearchScopes } from './SearchScopes'
