+++ web/src/search/queryBuilder/QueryBuilder.scss
