+++ web/src/search/queryBuilder/QueryBuilderInputRow.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { InfoDropdown } from './InfoDropdown'
+import { InfoDropdown } from '../input/InfoDropdown'
 import { QueryBuilderState } from './QueryBuilder'
 
 /** An example demonstrating the capabilities of the search field. */
