+++ web/src/search/queryBuilder/index.scss
@@ -1,0 +1,2 @@
+@import './QueryBuilder';
+@import './QueryBuilderInputRow';
